make parse_model_name parse 1.0

+++ src/utils.py
@@ -144,7 +144,7 @@ def parse_model_name(model_name, as_dict=False):
     model_name = Path(model_name).stem
     exp = r"([A-Za-z0-9]+)-([A-Za-z]+)([0-9]+)-([A-Za-z]*)-?([0-9]+[\.]?[0-9]*)x-([a-z]+)(?:.[A-Za-z]+)?"
     dataset, model_type, size, batch_norm, width, variant = re.match(exp, model_name).groups()
-    size, width = int(size), int(width)
+    size, width = float(size), float(width)
     batch_norm = "bn" in batch_norm
     if as_dict:
         return {
